requirements.txt
pandas>=1.5.0
pdfplumber>=0.7.6

google-generativeai>=0.3.0
groq>=0.3.0

python-dotenv>=1.0.0
pytest>=7.0.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
